Carry numeric and bigint values as exact decimal strings on every transport
A JSON number and a GraphQL Float are the same IEEE double: a numeric
column rounded and a bigint beyond 2^53 lost digits, while the schemas
claimed number and integer and the serializer turned only the safe bigints
into numbers. The scalar table now declares both as decimal strings with a
pattern and a description; the row serializer prints every numeric and
bigint value as text on every public boundary; GraphQL gains a Decimal
scalar that serializes text and accepts a number or a string; REST filters
on those columns keep the decimal text; the web generators map Decimal to a
number in forms and a string in records. A test serializes rows with the
driver's text, a handler's number and a bigint for every entity and
validates them against the REST and MCP record schemas.

Client-visible: a numeric or bigint field arrives as "12.50" rather than
12.5; inputs stay JSON numbers.

+++ packages/compiler/src/authoring/generators/pages.ts
@@ -1873,7 +1873,8 @@ function graphQlFieldTypeToFormValueType(type: string): GeneratedFormFieldConfig
     ID: "string",
     String: "string",
     Int: "integer",
-    Float: "number",
+    // A form edits a decimal as a number; the record it reads it back from carries a string.
+    Decimal: "number",
     Boolean: "boolean",
     JSON: "object",
   };
